requirements.txt
openai
tqdm
deepeval
prettytable
scipy
dify_client
jsonlines
python-dotenv
